Refactored code, allowed assessments to be imported along with SLOs, and improved the logic and changed the database regarding new vs imprted assessments and slos

+++ makeReports/views/assessment_views.py
@@ -12,34 +12,22 @@ from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
 from django.utils import timezone
 from django.views.generic.edit import FormMixin
 from makeReports.views.helperFunctions.section_context import *
+from makeReports.views.helperFunctions.mixins import *
 
-class AssessmentSummary(LoginRequiredMixin,UserPassesTestMixin,ListView):
+class AssessmentSummary(DeptReportMixin,ListView):
     model = AssessmentVersion
     template_name = "makeReports/Assessment/assessmentSummary.html"
     context_object_name = "assessment_list"
-    def dispatch(self,request,*args,**kwargs):
-        self.report = Report.objects.get(pk=self.kwargs['report'])
-        return super(AssessmentSummary,self).dispatch(request,*args,**kwargs)
     def get_queryset(self):
         report = self.report
         objs = AssessmentVersion.objects.filter(report=report).order_by("slo__number","number")
         return objs
     def get_context_data(self, **kwargs):
         context = super(AssessmentSummary, self).get_context_data()
-        context['rpt'] = self.report
         return section2Context(self,context)
-    def test_func(self):
-        return (self.report.degreeProgram.department == self.request.user.profile.department)
-class AddNewAssessment(LoginRequiredMixin,UserPassesTestMixin,FormView):
+class AddNewAssessment(DeptReportMixin,FormView):
     template_name = "makeReports/Assessment/addAssessment.html"
     form_class = CreateNewAssessment
-    def dispatch(self,request,*args,**kwargs):
-        self.report = Report.objects.get(pk=self.kwargs['report'])
-        return super(AddNewAssessment,self).dispatch(request,*args,**kwargs)
-    def get_context_data(self,**kwargs):
-        context = super(AddNewAssessment,self).get_context_data(**kwargs)
-        context['rpt'] = self.report
-        return context
     def get_form_kwargs(self):
         kwargs = super(AddNewAssessment,self).get_form_kwargs()
         kwargs['sloQS'] = SLOInReport.objects.filter(report=self.report).order_by("number")
@@ -50,8 +38,27 @@ class AddNewAssessment(LoginRequiredMixin,UserPassesTestMixin,FormView):
         rpt = self.report
         form.cleaned_data['slo'].numberOfAssess += 1
         form.cleaned_data['slo'].save()
-        assessObj = Assessment.objects.create(title=form.cleaned_data['title'], domainExamination=False, domainProduct=False, domainPerformance=False, directMeasure =form.cleaned_data['directMeasure'])
-        assessRpt = AssessmentVersion.objects.create(date=datetime.now(), number=form.cleaned_data['slo'].numberOfAssess, assessment=assessObj, description=form.cleaned_data['description'], finalTerm=form.cleaned_data['finalTerm'], where=form.cleaned_data['where'], allStudents=form.cleaned_data['allStudents'], sampleDescription=form.cleaned_data['sampleDescription'], frequency=form.cleaned_data['frequency'], threshold=form.cleaned_data['threshold'], target=form.cleaned_data['target'] ,slo=form.cleaned_data['slo'] ,firstInstance= True, report=rpt, changedFromPrior=False)
+        assessObj = Assessment.objects.create(
+            title=form.cleaned_data['title'], 
+            domainExamination=False, 
+            domainProduct=False, 
+            domainPerformance=False, 
+            directMeasure =form.cleaned_data['directMeasure'])
+        assessRpt = AssessmentVersion.objects.create(
+            date=datetime.now(), 
+            number=form.cleaned_data['slo'].numberOfAssess, 
+            assessment=assessObj, 
+            description=form.cleaned_data['description'], 
+            finalTerm=form.cleaned_data['finalTerm'], 
+            where=form.cleaned_data['where'], 
+            allStudents=form.cleaned_data['allStudents'], 
+            sampleDescription=form.cleaned_data['sampleDescription'], 
+            frequency=form.cleaned_data['frequency'], 
+            threshold=form.cleaned_data['threshold'], 
+            target=form.cleaned_data['target'],
+            slo=form.cleaned_data['slo'],
+            report=rpt, 
+            changedFromPrior=False)
         dom = form.cleaned_data['domain']
         if ("Pe" in dom):
             assessObj.domainPerformance = True
@@ -62,8 +69,6 @@ class AddNewAssessment(LoginRequiredMixin,UserPassesTestMixin,FormView):
         assessObj.save()
         assessRpt.save()
         return super(AddNewAssessment, self).form_valid(form)
-    def test_func(self):
-        return (self.report.degreeProgram.department == self.request.user.profile.department)
 class AddNewAssessmentSLO(AddNewAssessment):
     def get_initial(self):
         initial = super(AddNewAssessmentSLO,self).get_initial()
@@ -71,12 +76,9 @@ class AddNewAssessmentSLO(AddNewAssessment):
         return initial
     def get_success_url(self):
         return reverse_lazy('makeReports:slo-summary', args=[self.report.pk])
-class ImportAssessment(LoginRequiredMixin,UserPassesTestMixin,FormView):
+class ImportAssessment(DeptReportMixin,FormView):
     template_name = "makeReports/Assessment/importAssessment.html"
     form_class = ImportAssessmentForm
-    def dispatch(self,request,*args,**kwargs):
-        self.report = Report.objects.get(pk=self.kwargs['report'])
-        return super(ImportAssessment,self).dispatch(request,*args,**kwargs)
     def get_success_url(self):
         return reverse_lazy('makeReports:assessment-summary', args=[self.report.pk])
     def get_form_kwargs(self):
@@ -105,7 +107,23 @@ class ImportAssessment(LoginRequiredMixin,UserPassesTestMixin,FormView):
         num = slo.numberOfAssess
         for assessVers in form.cleaned_data['assessment']:
             num += 1
-            AssessmentVersion.objects.create(slo=slo,number=num,date=datetime.now(), description=assessVers.description ,assessment=assessVers.assessment, firstInstance=False, report=rpt, changedFromPrior=False, finalTerm=assessVers.finalTerm, where=assessVers.where, allStudents=assessVers.allStudents, sampleDescription=assessVers.sampleDescription, frequency=assessVers.frequency, threshold=assessVers.threshold, target=assessVers.target)
+            AssessmentVersion.objects.create(
+                slo=slo,
+                number=num,
+                date=datetime.now(), 
+                description=assessVers.description,
+                assessment=assessVers.assessment, 
+                report=rpt, 
+                changedFromPrior=False, 
+                finalTerm=assessVers.finalTerm, 
+                where=assessVers.where, 
+                allStudents=assessVers.allStudents, 
+                sampleDescription=assessVers.sampleDescription, 
+                frequency=assessVers.frequency, 
+                threshold=assessVers.threshold, 
+                target=assessVers.target)
+            assessVers.assessment.numberOfUses += 1
+            assessVers.assessment.save()
         slo.numberOfAssess = num
         slo.save()
         return super(ImportAssessment,self).form_valid(form)
@@ -115,10 +133,7 @@ class ImportAssessment(LoginRequiredMixin,UserPassesTestMixin,FormView):
         context['currentDPpk'] = r.degreeProgram.pk
         context['degPro_list'] = DegreeProgram.objects.filter(department=r.degreeProgram.department)
         context['slo_list'] = SLOInReport.objects.filter(report=r).order_by("number").union(SLOInReport.objects.filter(report__degreeProgram__department=r.degreeProgram.department).exclude(report=r).order_by("number"),all=True)
-        context['rpt']= r
         return context
-    def test_func(self):
-        return (self.report.degreeProgram.department == self.request.user.profile.department)
 class ImportAssessmentSLO(ImportAssessment):
     template_name = "makeReports/Assessment/importAssessmentSLO.html"
     def dispatch(self,request,*args,**kwargs):
@@ -134,11 +149,10 @@ class ImportAssessmentSLO(ImportAssessment):
         context = super(ImportAssessmentSLO, self).get_context_data(**kwargs)
         context['slo'] = self.slo
         return context
-class EditImportedAssessment(LoginRequiredMixin,UserPassesTestMixin,FormView):
+class EditImportedAssessment(DeptReportMixin,FormView):
     template_name = "makeReports/Assessment/editImportedAssessment.html"
     form_class = EditImportedAssessmentForm
     def dispatch(self,request,*args,**kwargs):
-        self.report = Report.objects.get(pk=self.kwargs['report'])
         self.assessVers = AssessmentVersion.objects.get(pk=self.kwargs['assessIR'])
         return super(EditImportedAssessment,self).dispatch(request,*args,**kwargs)
     def get_initial(self):
@@ -157,16 +171,10 @@ class EditImportedAssessment(LoginRequiredMixin,UserPassesTestMixin,FormView):
         kwargs = super(EditImportedAssessment,self).get_form_kwargs()
         kwargs['sloQS'] = SLOInReport.objects.filter(report=self.report).order_by("number")
         return kwargs
-    def get_context_data(self, **kwargs):
-        context = super(EditImportedAssessment,self).get_context_data(**kwargs)
-        context['rpt'] = self.report
-        return context
     def get_success_url(self):
         r = Report.objects.get(pk=self.kwargs['report'])
         return reverse_lazy('makeReports:assessment-summary', args=[r.pk])
     def form_valid(self,form):
-        #assessVers = AssessmentVersion.objects.get(pk=self.request.GET['assessIR'])
-        #newAssessVers = AssessmentVersion.objects.create(date=datetime.now(), description=form.cleaned_data['description'], assessment=assessVers.assessment, changedFromPrior=False, firstInstance=False, finalTerm=form.cleaned_data['finalTerm'], where=form.cleaned_data['where'], allStudents=form.cleaned_data['allStudents'], sampleDescription=form.cleaned_data['sampleDescription'], frequency=form.cleaned_data['frequency'], threshold=form.cleaned_data['threshold'], target=form.cleaned_data['target'])
         r = self.report
         self.assessVers.description = form.cleaned_data['description']
         self.assessVers.date = datetime.now()
@@ -181,13 +189,10 @@ class EditImportedAssessment(LoginRequiredMixin,UserPassesTestMixin,FormView):
         self.assessVers.save()
         self.assessVers.assessment.save()
         return super(EditImportedAssessment, self).form_valid(form)
-    def test_func(self):
-        return (self.report.degreeProgram.department == self.request.user.profile.department)
-class EditNewAssessment(LoginRequiredMixin,UserPassesTestMixin,FormView):
+class EditNewAssessment(DeptReportMixin,FormView):
     template_name = "makeReports/Assessment/editNewAssessment.html"
     form_class = EditNewAssessmentForm
     def dispatch(self,request,*args,**kwargs):
-        self.report = Report.objects.get(pk=self.kwargs['report'])
         self.assessVers = AssessmentVersion.objects.get(pk=self.kwargs['assessIR'])
         return super(EditNewAssessment,self).dispatch(request,*args,**kwargs)
     def get_initial(self):
@@ -207,10 +212,6 @@ class EditNewAssessment(LoginRequiredMixin,UserPassesTestMixin,FormView):
         initial['target'] = self.assessVers.target
         initial['slo'] = self.assessVers.slo
         return initial
-    def get_context_data(self,**kwargs):
-        context = super(EditNewAssessment,self).get_context_data(**kwargs)
-        context['rpt'] = self.report
-        return context
     def get_form_kwargs(self):
         kwargs = super(EditNewAssessment,self).get_form_kwargs()
         kwargs['sloQS'] = SLOInReport.objects.filter(report=self.report).order_by("number")
@@ -234,14 +235,11 @@ class EditNewAssessment(LoginRequiredMixin,UserPassesTestMixin,FormView):
         self.assessVers.save()
         self.assessVers.assessment.save()
         return super(EditNewAssessment,self).form_valid(form)
-    def test_func(self):
-        return (self.report.degreeProgram.department == self.request.user.profile.department)
-class SupplementUpload(LoginRequiredMixin,UserPassesTestMixin,CreateView):
+class SupplementUpload(DeptReportMixin,CreateView):
     template_name = "makeReports/Assessment/supplementUpload.html"
     model = AssessmentSupplement
     fields = ['supplement']
     def dispatch(self,request,*args,**kwargs):
-        self.report = Report.objects.get(pk=self.kwargs['report'])
         self.assessVers = AssessmentVersion.objects.get(pk=self.kwargs['assessIR'])
         return super(SupplementUpload,self).dispatch(request,*args,**kwargs)
     def get_success_url(self):
@@ -252,17 +250,10 @@ class SupplementUpload(LoginRequiredMixin,UserPassesTestMixin,CreateView):
         form.instance.assessmentVersion = self.assessVers
         form.instance.uploaded_at = datetime.now()
         return super(SupplementUpload,self).form_valid(form)
-    def get_context_data(self, **kwargs):
-        context = super(SupplementUpload, self).get_context_data(**kwargs)
-        context['rpt']=self.report
-        return context
-    def test_func(self):
-        return (self.report.degreeProgram.department==self.request.user.profile.department)
-class ImportSupplement(LoginRequiredMixin,UserPassesTestMixin,FormView):
+class ImportSupplement(DeptReportMixin,FormView):
     template_name = "makeReports/Assessment/importSupplement.html"
     form_class = ImportSupplementsForm
     def dispatch(self,request,*args,**kwargs):
-        self.report = Report.objects.get(pk=self.kwargs['report'])
         self.assessVers = AssessmentVersion.objects.get(pk=self.kwargs['assessIR'])
         return super(ImportSupplement,self).dispatch(request,*args,**kwargs)
     def get_success_url(self):
@@ -271,7 +262,8 @@ class ImportSupplement(LoginRequiredMixin,UserPassesTestMixin,FormView):
          kwargs = super(ImportSupplement,self).get_form_kwargs()
          yearIn = self.request.GET['year']
          dPobj = DegreeProgram.objects.get(pk=self.request.GET['dp'])
-         kwargs['supChoices'] = AssessmentSupplement.objects.filter(report__year=yearIn, report__degreeProgram=dPobj)
+         kwargs['supChoices'] = AssessmentSupplement.objects.filter(
+             report__year=yearIn, report__degreeProgram=dPobj)
          return kwargs
     def form_valid(self,form):
         self.assessVers.supplements.add(form.cleaned_data['sup'])
@@ -281,59 +273,34 @@ class ImportSupplement(LoginRequiredMixin,UserPassesTestMixin,FormView):
         context = super(ImportSupplement, self).get_context_data(**kwargs)
         context['currentDPpk'] = self.report.degreeProgram.pk
         context['degPro_list'] = DegreeProgram.objects.filter(department=self.report.degreeProgram.department)
-        context['rpt'] = self.report
         return context
-    def test_func(self):
-        return (self.report.degreeProgram.department == self.request.user.profile.department)
-class DeleteSupplement(LoginRequiredMixin,UserPassesTestMixin,DeleteView):
+class DeleteSupplement(DeptReportMixin,DeleteView):
     model = AssessmentSupplement
     template_name = "makeReports/Assessment/deleteSupplement.html"
-    def dispatch(self,request,*args,**kwargs):
-        self.report = Report.objects.get(pk=self.kwargs['report'])
-        return super(DeleteSupplement,self).dispatch(request,*args,**kwargs)
-    def get_context_data(self,**kwargs):
-        context = super(DeleteSupplement,self).get_context_data(**kwargs)
-        context['rpt'] = self.report
-        return context
     def get_success_url(self):
         return reverse_lazy('makeReports:assessment-summary', args=[self.report.pk])
-    def test_func(self):
-        return (self.report.degreeProgram.department == self.request.user.profile.department)
-class Section2Comment(LoginRequiredMixin,UserPassesTestMixin,FormView):
+class Section2Comment(DeptReportMixin,FormView):
     template_name = "makeReports/Assessment/assessmentComment.html"
     form_class = Single2000Textbox
-    def dispatch(self,request,*args,**kwargs):
-        self.report = Report.objects.get(pk=self.kwargs['report'])
-        return super(Section2Comment,self).dispatch(request,*args,**kwargs)
     def get_success_url(self):
         return reverse_lazy('makeReports:assessment-summary', args=[self.report.pk])
     def form_valid(self, form):
         self.report.section2Comment = form.cleaned_data['text']
         self.report.save()
         return super(Section2Comment,self).form_valid(form)
-    def get_context_data(self,**kwargs):
-        context = super(Section2Comment,self).get_context_data(**kwargs)
-        context['rpt'] = self.report
-        return context
     def get_initial(self):
         initial = super(Section2Comment,self).get_initial()
         initial['text']="No comment."
         return initial
-    def test_func(self):
-        return (self.report.degreeProgram.department == self.request.user.profile.department)
-class DeleteImportedAssessment(DeleteView):
+class DeleteImportedAssessment(DeptReportMixin,DeleteView):
     model = AssessmentVersion
     template_name = "makeReports/Assessment/deleteAssessment.html"
     def dispatch(self,request,*args,**kwargs):
-        self.report = Report.objects.get(pk=self.kwargs['report'])
         a = AssessmentVersion.objects.get(pk=self.kwargs['pk'])
         self.oldNum = a.number
         self.slo = a.slo
+        self.assessment = a.assessment
         return super(DeleteImportedAssessment,self).dispatch(request,*args,**kwargs)
-    def get_context_data(self,**kwargs):
-        context = super(DeleteImportedAssessment,self).get_context_data(**kwargs)
-        context['rpt'] = self.report
-        return context
     def get_success_url(self):
         oldNum = self.oldNum
         num = self.slo.numberOfAssess
@@ -342,14 +309,19 @@ class DeleteImportedAssessment(DeleteView):
             if a.number > oldNum:
                 a.number -= 1
                 a.save()
+        if self.assessment.numberOfUses <= 1:
+            self.assessment.delete()
+            self.assessment.save()
+        else:
+            self.assessment.numberOfUses -= 1
+            self.assessment.save()
         self.slo.numberOfAssess -= 1
         self.slo.save()
         return reverse_lazy('makeReports:assessment-summary', args=[self.report.pk])
-class DeleteNewAssessment(DeleteView):
+class DeleteNewAssessment(DeptReportMixin,DeleteView):
     model = AssessmentVersion
     template_name = "makeReports/Assessment/deleteAssessment.html"
     def dispatch(self,request,*args,**kwargs):
-        self.report = Report.objects.get(pk=self.kwargs['report'])
         a = AssessmentVersion.objects.get(pk=self.kwargs['pk'])
         self.oldNum = a.number
         self.slo = a.slo
@@ -357,8 +329,12 @@ class DeleteNewAssessment(DeleteView):
     def get_success_url(self):
         ASSESSIR = AssessmentVersion.objects.get(pk=self.kwargs['pk'])
         assessment = ASSESSIR.assessment
-        assessment.delete()
-        assessment.save()
+        if assessment.numberOfUses <= 1:
+            assessment.delete()
+            assessment.save()
+        else:
+            assessment.numberOfUses -= 1
+            assessment.save()
         oldNum = self.oldNum
         num = self.slo.numberOfAssess
         assess = AssessmentVersion.objects.filter(report=self.report).order_by("slo__number","number")
@@ -369,7 +345,3 @@ class DeleteNewAssessment(DeleteView):
         self.slo.numberOfAssess -= 1
         self.slo.save()        
         return reverse_lazy('makeReports:assessment-summary', args=[self.report.pk])
-    def get_context_data(self,**kwargs):
-        context = super(DeleteNewAssessment,self).get_context_data(**kwargs)
-        context['rpt'] = self.report
-        return context
